+++ crates/sui-core/src/authority.rs
@@ -17,7 +17,6 @@ use prometheus::{
     register_histogram_with_registry, register_int_counter_with_registry,
     register_int_gauge_with_registry, Histogram, IntCounter, IntGauge, Registry,
 };
-use std::cmp::Ordering as CmpOrdering;
 use std::path::PathBuf;
 use std::str::FromStr;
 use std::time::Duration;
@@ -2167,58 +2166,6 @@ impl AuthorityState {
         self.database.get_latest_parent_entry(object_id)
     }
 
-    /// Verifies transaction signatures and other data
-    /// Important: This function can potentially be called in parallel and you can not rely on order of transactions to perform verification
-    /// If this function return an error, transaction is skipped and is not passed to handle_consensus_transaction
-    /// This function returns unit error and is responsible for emitting log messages for internal errors
-    pub(crate) fn verify_consensus_transaction(
-        &self,
-        current_epoch: EpochId,
-        transaction: SequencedConsensusTransaction,
-    ) -> Result<VerifiedSequencedConsensusTransaction, ()> {
-        let _scope = monitored_scope("VerifyConsensusTransaction");
-        if self
-            .load_epoch_store(current_epoch)
-            .map_err(|err| {
-                debug!(
-                    "Error loading epoch store in verify_consensus_transaction: {:?}",
-                    err
-                )
-            })?
-            .is_consensus_message_processed(&transaction.transaction.key())
-            .expect("Storage error")
-        {
-            debug!(
-                consensus_index=?transaction.consensus_index.index.transaction_index,
-                tracking_id=?transaction.transaction.tracking_id,
-                "handle_consensus_transaction UserTransaction [skip]",
-            );
-            self.metrics.skipped_consensus_txns.inc();
-            return Err(());
-        }
-        // Signatures are verified as part of narwhal payload verification in SuiTxValidator
-        match &transaction.transaction.kind {
-            ConsensusTransactionKind::UserTransaction(_certificate) => {}
-            ConsensusTransactionKind::CheckpointSignature(data) => {
-                if transaction.sender_authority() != data.summary.auth_signature.authority {
-                    warn!("CheckpointSignature authority {} does not match narwhal certificate source {}", data.summary.auth_signature.authority, transaction.certificate.origin() );
-                    return Err(());
-                }
-            }
-            ConsensusTransactionKind::EndOfPublish(authority) => {
-                if &transaction.sender_authority() != authority {
-                    warn!(
-                        "EndOfPublish authority {} does not match narwhal certificate source {}",
-                        authority,
-                        transaction.certificate.origin()
-                    );
-                    return Err(());
-                }
-            }
-        }
-        Ok(VerifiedSequencedConsensusTransaction(transaction))
-    }
-
     /// The transaction passed here went through verification in verify_consensus_transaction.
     /// This method is called in the exact sequence message are ordered in consensus.
     /// Errors returned by this call are treated as critical errors and cause node to panic.
@@ -2343,40 +2290,6 @@ impl AuthorityState {
                 Ok(None)
             }
         }
-    }
-
-    pub(crate) fn handle_commit_boundary<C: CheckpointServiceNotify>(
-        &self,
-        current_epoch: EpochId,
-        committed_dag: &Arc<CommittedSubDag>,
-        checkpoint_service: &Arc<C>,
-    ) -> SuiResult {
-        let round = committed_dag.round();
-        debug!("Commit boundary at {}", round);
-        // This exchange is restart safe because of following:
-        //
-        // We try to read last checkpoint content and send it to the checkpoint service
-        // CheckpointService::notify_checkpoint is idempotent in case you send same last checkpoint multiple times
-        //
-        // Only after CheckpointService::notify_checkpoint stores checkpoint in it's store we update checkpoint boundary
-        let epoch_store = self.load_epoch_store(current_epoch)?;
-        if let Some((index, roots)) = epoch_store.last_checkpoint(round)? {
-            let final_checkpoint_round = epoch_store.final_epoch_checkpoint()?;
-            let final_checkpoint = match final_checkpoint_round.map(|r| r.cmp(&round)) {
-                Some(CmpOrdering::Less) => {
-                    debug!(
-                        "Not forming checkpoint for round {} above final checkpoint round {:?}",
-                        round, final_checkpoint_round
-                    );
-                    return Ok(());
-                }
-                Some(CmpOrdering::Equal) => true,
-                Some(CmpOrdering::Greater) => false,
-                None => false,
-            };
-            checkpoint_service.notify_checkpoint(index, roots, final_checkpoint)?;
-        }
-        epoch_store.record_checkpoint_boundary(round)
     }
 
     pub async fn create_advance_epoch_tx_cert(
